add the MainActivity and music list
