+++ app/src/main/java/com/sample/andremion/musicplayer/activities/MusicListActivity.java
@@ -237,18 +237,6 @@ public class MusicListActivity extends PlayActivity {
         }
     }
 
-    /*
-    当用户回退时保证Activity不销毁
-     */
-    @Override
-    public boolean onKeyDown(int keyCode, KeyEvent event) {
-        if(keyCode == KeyEvent.KEYCODE_BACK){
-            moveTaskToBack(true);
-            return true;
-        }
-        return super.onKeyDown(keyCode, event);
-    }
-
     @Override
     protected void onDestroy() {
         super.onDestroy();
