+++ app/src/main/java/com/sample/andremion/musicplayer/adapter/MusicListAlbumAdapter.java
@@ -1,31 +1,25 @@
-package com.sample.andremion.musicplayer.view;
+package com.sample.andremion.musicplayer.adapter;
 
-import android.content.ComponentName;
 import android.content.Context;
 import android.content.Intent;
-import android.support.v4.content.LocalBroadcastManager;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.AdapterView;
 import android.widget.ImageView;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import com.sample.andremion.musicplayer.R;
-import com.sample.andremion.musicplayer.activities.MainActivity;
-import com.sample.andremion.musicplayer.activities.Music;
+import com.sample.andremion.musicplayer.model.MusicListAlbum;
 import com.sample.andremion.musicplayer.activities.MusicListActivity;
-import com.sample.andremion.musicplayer.music.PlayService;
 
 import java.util.List;
 
-public class MusicListAdapter extends RecyclerView.Adapter<MusicListAdapter.ViewHolder> {
-    private final List<Music> myMusicList;
+public class MusicListAlbumAdapter extends RecyclerView.Adapter<MusicListAlbumAdapter.ViewHolder> {
+    private final List<MusicListAlbum> myMusicList;
     private Context mContext;
    // private final LocalBroadcastManager broadcastManager;
-    public MusicListAdapter(List<Music> myMusicList, Context context){
+    public MusicListAlbumAdapter(List<MusicListAlbum> myMusicList, Context context){
         this.myMusicList = myMusicList;
         mContext = context;
     }
@@ -56,7 +50,7 @@ public class MusicListAdapter extends RecyclerView.Adapter<MusicListAdapter.View
             @Override
             public void onClick(View v) {
                 int position =holder.getAdapterPosition();
-                Music music=myMusicList.get(position);
+                MusicListAlbum music=myMusicList.get(position);
                 //Toast.makeText(v.getContext(),"sdadasdddsad", Toast.LENGTH_SHORT).show();
                 Intent intent = new Intent(mContext,MusicListActivity.class);
                 //  ComponentName componentName = new ComponentName("com.sample.andremion.musicplayer.view", "com.sample.andremion.musicplayer.v");
@@ -69,10 +63,10 @@ public class MusicListAdapter extends RecyclerView.Adapter<MusicListAdapter.View
 
     @Override
     public void onBindViewHolder(final ViewHolder holder,final int position){
-        Music music = myMusicList.get(position);
+        MusicListAlbum music = myMusicList.get(position);
         holder.musicImage.setImageResource(music.getImageId());
         holder.musicName.setText(music.getName());
-        holder.musicName.setText(music.getNumber());
+        //holder.musicName.setText(music.getNumber());
 
     }
